travel_project/requirements.txt
asgiref==3.5.2
attrs==21.4.0
backports.zoneinfo==0.2.1
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-auth==0.1.0
importlib-metadata==4.11.4
Markdown==3.3.7
more-itertools==8.13.0
packaging==21.3
pip==22.0.3
pluggy==0.13.1
py==1.11.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==5.4.3
pytz==2022.1
setuptools==60.6.0
sqlparse==0.4.2
wcwidth==0.2.5
wheel==0.37.1
zipp==3.8.0
